Create tests for TxStateManager
